selenium-setup/requirements.txt
selenium
cryptography
requests
gspread
google-auth
